requirements.txt
streamlit==1.32.0
pandas==2.2.0
plotly==5.18.0
pyyaml==6.0.1
openpyxl==3.1.2
unidecode==1.3.7 